Abstract Travis job updates
So we can re-use it for other CI services
